+++ ForecastRestWithLibs/app/src/main/java/com/android2ee/formation/restservice/forecastyahoo/withlibs/transverse/event/WeatherDownloadedEvent.java
@@ -39,9 +39,9 @@ import com.android2ee.formation.restservice.forecastyahoo.withlibs.transverse.mo
 @Deprecated
 public class WeatherDownloadedEvent {
     Weather weather;
-    int cityId;
+    Long cityId;
 
-    public WeatherDownloadedEvent(Weather weather,int cityId) {
+    public WeatherDownloadedEvent(Weather weather,Long cityId) {
         this.cityId=cityId;
         this.weather = weather;
     }
@@ -57,11 +57,11 @@ public class WeatherDownloadedEvent {
         this.weather = weather;
     }
 
-    public int getCityId() {
+    public Long getCityId() {
         return cityId;
     }
 
-    public void setCityId(int cityId) {
+    public void setCityId(Long cityId) {
         this.cityId = cityId;
     }
 }
